Update styles to work with theme-ui v0.8
