+++ starters/gatsby-starter-catalyst-blog/src/gatsby-plugin-theme-ui/index.js
@@ -71,11 +71,6 @@ const theme = merge(BaseTheme, {
       },
     },
   },
-  variants: {
-    siteTitle: {
-      fontSize: [3, 4, null, 5, null],
-    },
-  },
 })
 
 export default theme
