+++ starters/gatsby-starter-catalyst-lithium/src/gatsby-plugin-theme-ui/index.js
@@ -73,7 +73,7 @@ const theme = merge(BaseTheme, {
   sizes: {
     ...tailwind.sizes,
     maxPageWidth: "1440px", // Sets the max width of elements like the header/footer large screens
-    maxContentWidth: "60ch", // Sets the container size on larger screens, e.g. tablets and laptops
+    maxContentWidth: "720px", // Sets the container size on larger screens, e.g. tablets and laptops
     iconsFooter: "32px", // Sets the icons size for the footer
     iconsHeader: "20px", // Sets the icons size for the header
   },
@@ -98,12 +98,6 @@ const theme = merge(BaseTheme, {
     },
   },
   variants: {
-    siteTitle: {
-      fontSize: [3, 4, null, 5, null],
-      "::after": {
-        content: "none",
-      },
-    },
     navLink: {
       fontFamily: "heading",
       fontWeight: "400",
