+++ themes/gatsby-theme-catalyst-helium/src/components/post-footer.js
@@ -25,7 +25,12 @@ const PostFooter = ({ previous, next }) => (
           {previous && (
             <div>
               <Themed.h5>&larr; Previous Post</Themed.h5>
-              <Themed.a as={Link} to={previous.slug} rel="prev">
+              <Themed.a
+                as={Link}
+                to={previous.slug}
+                rel="prev"
+                sx={{ fontSize: 2 }}
+              >
                 {previous.title}
               </Themed.a>
             </div>
@@ -44,7 +49,12 @@ const PostFooter = ({ previous, next }) => (
               }}
             >
               <Themed.h5>Next Post &rarr;</Themed.h5>
-              <Themed.a as={Link} to={next.slug} rel="next">
+              <Themed.a
+                as={Link}
+                to={next.slug}
+                rel="next"
+                sx={{ fontSize: 2 }}
+              >
                 {next.title}
               </Themed.a>
             </div>
