--- themes/gatsby-theme-catalyst-hydrogen/src/gatsby-theme-catalyst-sanity/components/sanity/sanity-base-theme.js
@@ -1,229 +1,0 @@
-// See https://theme-ui.com/ for more info and also https://www.gatsbyjs.com/docs/theme-ui/
-// Try changing some of the colors below to see what happens.
-import { tailwind, baseColors } from "@theme-ui/preset-tailwind"
-import prism from "@theme-ui/prism/presets/oceanic-next"
-
-export default {
-  ...tailwind,
-  breakpoints: ["480px", "768px", "1024px", "1440px"],
-  fonts: {
-    ...tailwind.fonts,
-    body:
-      'system-ui,-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,"Helvetica Neue",Arial,"Noto Sans",sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol","Noto Color Emoji"',
-    heading: "inherit",
-    monospace:
-      'Menlo,Monaco,Consolas,"Liberation Mono","Courier New",monospace',
-  },
-  colors: {
-    ...tailwind.colors,
-    background: baseColors.gray[1], //Try "#954264",
-    text: baseColors.gray[8],
-    primary: baseColors.red[8],
-    secondary: baseColors.red[5],
-    accent: baseColors.gray[2],
-    highlight: baseColors.orange[5],
-    muted: baseColors.gray[2],
-    header: {
-      background: "transparent",
-      backgroundOpen: baseColors.red[1],
-      text: baseColors.gray[8],
-      textOpen: baseColors.gray[8],
-      icons: baseColors.gray[6],
-      iconsHover: baseColors.red[8],
-      iconsOpen: baseColors.gray[6],
-    },
-    footer: {
-      background: baseColors.gray[2],
-      text: baseColors.gray[8],
-      links: baseColors.gray[8],
-      icons: baseColors.gray[8],
-    },
-  },
-  sizes: {
-    ...tailwind.sizes,
-    maxPageWidth: "1440px", // Sets the max width of elements like the header/footer large screens
-    maxContentWidth: "60ch", // Sets the container size on larger screens, e.g. tablets and laptops
-    headerHeight: "auto", // Provides fallback setting to control header height
-    logoWidthXS: "40px", // Logo width on extra small screens, up to 480px
-    logoWidthS: "50px", // Logo width on small screens, 480px - 768px
-    logoWidthM: "50px", // Logo width on medium screens, 768px - 1024px
-    logoWidthL: "60px", // Logo width on large screens, 1024px - 1440px
-    logoWidthXL: "60px", // Logo width on extra large screens, above 1440px
-    logoHeightXS: "40px", // Logo height on extra small screens, up to 480px
-    logoHeightS: "50px", // Logo height on small screens, 480px - 768px
-    logoHeightM: "50px", // Logo height on medium screens, 768px - 1024px
-    logoHeightL: "60px", // Logo height on large screens, 1024px - 1440px
-    logoHeightXL: "60px", // Logo height on extra large screens, above 1440px
-    iconsFooter: "32px", // Sets the icons size for the footer
-    iconsHeader: "20px", // Sets the icons size for the header
-  },
-  styles: {
-    ...tailwind.styles,
-    root: {
-      backgroundColor: "background",
-      color: "text",
-      fontFamily: "body",
-      fontWeight: "body",
-      lineHeight: "body",
-      fontSize: 2,
-    },
-    a: {
-      color: "primary",
-      textDecoration: "none",
-      ":hover": {
-        textDecoration: "underline",
-      },
-    },
-    h1: {
-      fontFamily: "heading",
-      fontWeight: "heading",
-      lineHeight: "heading",
-      m: 0,
-      mb: 1,
-      fontSize: 6,
-      mt: 4,
-    },
-    h2: {
-      fontFamily: "heading",
-      fontWeight: "heading",
-      lineHeight: "heading",
-      m: 0,
-      mb: 1,
-      fontSize: 5,
-      mt: 4,
-    },
-    h3: {
-      fontFamily: "heading",
-      fontWeight: "heading",
-      lineHeight: "heading",
-      m: 0,
-      mb: 1,
-      fontSize: 4,
-      mt: 3,
-    },
-    h4: {
-      fontFamily: "heading",
-      fontWeight: "heading",
-      lineHeight: "heading",
-      m: 0,
-      mb: 1,
-      fontSize: 3,
-    },
-    h5: {
-      fontFamily: "heading",
-      fontWeight: "heading",
-      lineHeight: "heading",
-      m: 0,
-      mb: 1,
-      fontSize: 2,
-    },
-    h6: {
-      fontFamily: "heading",
-      fontWeight: "heading",
-      lineHeight: "heading",
-      m: 0,
-      mb: 2,
-      fontSize: 1,
-    },
-    blockquote: {
-      bg: "muted",
-      p: 3,
-      borderLeft: "5px solid",
-      borderColor: "primary",
-    },
-    inlineCode: {
-      color: "text",
-      bg: "muted",
-      fontSize: 1,
-      p: 1,
-    },
-    pre: {
-      ...prism,
-      fontSize: 1,
-      p: 3,
-    },
-    table: {
-      width: "100%",
-      my: 3,
-      borderCollapse: "collapse",
-    },
-    th: {
-      verticalAlign: "bottom",
-      borderWidth: "2px",
-      borderStyle: "solid",
-      borderColor: "muted",
-      backgroundColor: "muted",
-      padding: 2,
-      textAlign: "inherit",
-    },
-    td: {
-      borderWidth: "2px",
-      borderStyle: "solid",
-      borderColor: "muted",
-      verticalAlign: "top",
-      padding: 2,
-    },
-  },
-  variants: {
-    siteTitle: {
-      fontSize: [4, null, null, 5, null],
-      "::after": {
-        display: "none",
-      },
-    },
-    navLink: {
-      "::after": {
-        position: "absolute",
-        top: "100%",
-        left: "0",
-        width: "100%",
-        height: "1px",
-        backgroundColor: "primary",
-        content: "''",
-        opacity: "0",
-        transition: "height 0.3s, opacity 0.3s, transform 0.3s",
-        transform: "translateY(-6px)",
-      },
-      ":hover, :focus, :active": {
-        textDecoration: "none",
-      },
-      ":hover::after, :active::after, :focus::after": {
-        height: "4px",
-        opacity: "1",
-        transform: "translateY(0px)",
-      },
-    },
-    navLinkActive: {
-      textDecoration: "none",
-      "::after": {
-        position: "absolute",
-        top: "100%",
-        left: "0",
-        width: "100%",
-        height: "4px",
-        backgroundColor: "primary",
-        content: "''",
-        opacity: "1",
-        transform: "translateY(0px)",
-      },
-    },
-    navLinkSub: {
-      "::after": {
-        content: "none",
-      },
-      ":hover::after": {
-        content: "none",
-      },
-      ":hover, :focus, :active": {
-        color: "primary",
-      },
-    },
-    navLinkSubActive: {
-      textDecoration: "none",
-      color: "primary",
-      "::after": {
-        content: "none",
-      },
-    },
-  },
-}
